Current Trends and Future Trends in Big Data Analytics
It is necessary to learn more about the technology behind big data and investigate important issues from the standpoint of information technology. In this paper, we examine the state of big data today, as well as its problems and new directions. We offer the main conclusions and the evolving technological scene.